retrieved the clients workouts from the clients side

+++ src/main/java/org/example/demo2/dao/ClientDAO.java
@@ -47,7 +47,7 @@ public class ClientDAO {
         }
     }
 
-    public Client findByUserId(Long userId) throws SQLException {
+    public Client findByUserId(int userId) throws SQLException {
         Connection connection = null;
         try {
             connection = dbConnection.getConnection();
